+++ src/contrib/google/sheets/ui/AppendSpreadsheetOptions.tsx
@@ -36,7 +36,7 @@ import { useOnChangeEffect } from "@/contrib/google/sheets/core/useOnChangeEffec
 import { requireGoogleHOC } from "@/contrib/google/sheets/ui/RequireGoogleApi";
 import { type Schema } from "@/types/schemaTypes";
 import { isExpression, isTemplateExpression } from "@/utils/expressionUtils";
-import { type SpreadsheetTarget } from "@/contrib/google/sheets/core/handlers";
+import { type SpreadsheetTarget } from "@/contrib/google/sheets/core/sheetsApi";
 
 const DEFAULT_FIELDS_SCHEMA: Schema = {
   type: "object",
